Added component for second player input and started the scene manager refactor

+++ 08241_ACW2/New_Lab1/Utility/RenderText.cs
@@ -48,7 +48,7 @@ namespace PongGame.Utility
             GL.BindTexture(TextureTarget.Texture2D, textTexture);
 
             BitmapData data = textBMP.LockBits(rect, ImageLockMode.ReadOnly, System.Drawing.Imaging.PixelFormat.Format32bppArgb);
-            GL.TexImage2D(TextureTarget.Texture2D, 0, PixelInternalFormat.Rgba, textBMP.Width, (int)textBMP.Height, 0, OpenTK.Graphics.OpenGL.PixelFormat.Bgra, PixelType.UnsignedByte, data.Scan0);
+            GL.TexImage2D(TextureTarget.Texture2D, 0, PixelInternalFormat.Rgba, textBMP.Width, textBMP.Height, 0, OpenTK.Graphics.OpenGL.PixelFormat.Bgra, PixelType.UnsignedByte, data.Scan0);
             textBMP.UnlockBits(data);
 
             GL.Begin(PrimitiveType.Quads);
